refactor: move context-menu and flow-category-renderer to utils layer
Move shared modules from src/components/ to src/utils/ to fix cross-layer
violations where 4 utils files were importing directly from components.
Update all import paths across the codebase.
